se agregan algunos repositorios y se testea endpoint de chofer

+++ src/main/kotlin/uberto/backendgrupo72025/Service/UsuarioService.kt
@@ -29,7 +29,7 @@ class UsuarioService(val repoUsuario: UsuarioRepository) {
 
     fun comentariosRecibidos(id :Long) = getUsuarioById(id).comentarios
 
-    fun calificacion(id :Long) = (puntajeTotal(id) / cantidadComentarios(id)).toDouble()
+    fun calificacion(id :Long) = puntajeTotal(id).toDouble() / cantidadComentarios(id).toDouble()
 
     fun puntajeTotal(id: Long) = comentariosRecibidos(id).sumOf { it.puntaje }
